backend/requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
chromadb
pandas
langchain-ollama
langchain-chroma